module4 Разбиение на модули
