+++ js/initData.js
@@ -1,0 +1,25 @@
+const PLACE = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+]
+const TIME = [
+  '12:00',
+  '13:00',
+  '14:00',
+]
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+]
+const PHOTOS = [
+  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
+]
+export {PLACE, TIME, FEATURES, PHOTOS}
